Move applyEffect method
