+++ src/main/scala/com/martomate/tripaint/model/CumulativeImageChange.scala
@@ -7,7 +7,7 @@ import scalafx.scene.paint.Color
 import scala.collection.mutable.ArrayBuffer
 
 class CumulativeImageChange {
-  private val pixelsChanged = ArrayBuffer.empty[(TriangleCoords, Color, Color)]
+  private val pixelsChanged = ArrayBuffer.empty[PixelChange]
 
   def done(description: String, image: ImageContent): ImageChange = {
     val change = new ImageChange(description, image, pixelsChanged.reverse.toVector)
@@ -15,7 +15,8 @@ class CumulativeImageChange {
     change
   }
 
-  def addChange(index: TriangleCoords, oldColor: Color, newColor: Color): Unit = addChange(change = (index, oldColor, newColor))
+  def addChange(index: TriangleCoords, oldColor: Color, newColor: Color): Unit =
+    addChange(PixelChange(index, oldColor, newColor))
 
-  def addChange(change: (TriangleCoords, Color, Color)): Unit = pixelsChanged += change
+  def addChange(change: PixelChange): Unit = pixelsChanged += change
 }
