+++ src/main/scala/com/martomate/tripaint/model/ImageChange.scala
@@ -5,19 +5,20 @@ import com.martomate.tripaint.model.coords.TriangleCoords
 import com.martomate.tripaint.model.undo.Change
 import scalafx.scene.paint.Color
 
-class ImageChange(val description: String, val image: ImageContent, pixelsChanged: Seq[(TriangleCoords, Color, Color)]) extends Change {
+class ImageChange(val description: String, val image: ImageContent, pixelsChanged: Seq[PixelChange]) extends Change {
   def redo(): Boolean = {
     val draw = image.storage
-    for ((index, _, newColor) <- pixelsChanged) draw(index) = newColor
+    for (ch <- pixelsChanged) draw(ch.coords) = ch.after
     image.changeTracker.tellListenersAboutBigChange()
     true
   }
 
   def undo(): Boolean = {
     val draw = image.storage
-    for ((index, oldColor, _) <- pixelsChanged) draw(index) = oldColor
+    for (ch <- pixelsChanged) draw(ch.coords) = ch.before
     image.changeTracker.tellListenersAboutBigChange()
     true
   }
 }
 
+case class PixelChange(coords: TriangleCoords, before: Color, after: Color)
